Remove sort_order from folder objects
Sort order is removed from folder objects so that the algorithm
evaluates a folder-level checksum based purely on content, i.e.
checksum only. The code is no-longer sensitive to renaming. This
should improve its portability and applicability. The version is
upped to v0.0.2 and remains 0.0.x until the algorithm is ironed
out some more.